+++ stack.go
@@ -8,8 +8,12 @@ import (
 	"regexp"
 	"runtime"
 	"strings"
+)
 
-	"golang.org/x/sys/unix"
+const (
+	compilerColor   = "blue"
+	thirdPartyColor = "yellow"
+	yourColor       = "red"
 )
 
 func stackp() {
@@ -17,19 +21,27 @@ func stackp() {
 	re := regexp.MustCompile(`\d:`) // this pattern is the one created in `readLastLine()`
 
 	traces := getStackTrace()
+	if len(traces) > 0 {
+		printWithColor(
+			fmt.Sprintf("LEGEND:\n compiler: %s\n thirdParty: %s\n yours: %s\n", compilerColor, thirdPartyColor, yourColor),
+			"DEFAULT",
+			true,
+		)
+	}
+
 	for _, v := range traces {
 		if strings.Contains(v, "go/src/") {
 			// compiler
-			printWithColor(v, "cyan", false)
+			printWithColor(v, compilerColor, false)
 		} else if goModCache != "" && strings.Contains(v, goModCache) {
 			// third party
-			printWithColor(v, "magenta", false)
+			printWithColor(v, thirdPartyColor, false)
 		} else if re.MatchString(v) {
 			// this is code snippets
-			printWithColor(v, "green", false)
+			printWithColor(v, yourColor, false)
 		} else {
 			// your code
-			printWithColor(v, "green", true)
+			printWithColor(v, yourColor, true)
 		}
 	}
 }
@@ -137,9 +149,9 @@ func setColor(code int, bold bool) {
 func printWithColor(s, color string, bold bool) {
 	// TODO: should be iota
 	colors := map[string]int{
-		// Go compiler == cyan
-		// Third party == magenta
-		// Your code == green
+		// Go compiler == compilerColor
+		// Third party == thirdPartyColor
+		// Your code == yourColor
 
 		"black":   30,
 		"red":     31,
@@ -176,7 +188,5 @@ func noColor() bool {
 		return true
 	}
 
-	fd := os.Stdout.Fd()
-	_, err := unix.IoctlGetTermios(int(fd), unix.TCGETS)
-	return err == nil
+	return false
 }