+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Most recent version is listed first.
 
 # v0.0.18
 - Add ability to diff items: https://github.com/komuw/kama/pull/67
+- Add legend to stack trace: https://github.com/komuw/kama/pull/68
 
 # v0.0.17
 - Add ability to dump items with circular references: https://github.com/komuw/kama/pull/66